Bug Fix, Unit test fixes
Previously, controller was unable to run functions when buttons were false, which would mean you lose half the functionality. Fixed by making every function pointer in the code a pair now, first being for when the function is false, the second being for when it is true.

This bug fix meant some of the unit tests needed to be redone, which they are now.