Update all other motions to use task-only
TL;DR, instead of a motion only spawning a task if it is async, a task
is *always* spawned. This prevents the issue where, if a synchronous
function is called from a competition task and the competition state
changed, the competition task (and synchronous function) is deleted. If
that motion function was holding `Chassis::mutex` when it was deleted,
any other motions will block on `Chassis::mutex` for TIMEOUT_MAX (which
is several days...)